requirements.txt
tree-sitter
tree-sitter-c-sharp
PyYAML
lxml
anthropic
click
